refactor(schemas): move media tables to schemas/media/
Per-dialect Drizzle table definitions for media, mediaFolders, and
imageSizes now live in schemas/media/{postgres,mysql,sqlite}.ts with a
dialect-aware barrel exposing mediaTables(dialect).

database/schema/<dialect>.ts re-exports from the new location so existing
imports and relations() blocks keep compiling. Relations themselves stay
in database/schema/<dialect>.ts during Plan A and consolidate in Task 17.
