+++ packages/nextly/src/database/schema/mysql.ts
@@ -318,111 +318,11 @@ export const activityLogRelations = relations(activityLog, ({ one }) => ({
   }),
 }));
 
-/**
- * Media table for storing uploaded files and images
- *
- * Supports various storage backends (Vercel Blob, S3, R2, local filesystem)
- * Stores file metadata in database, actual files in configured storage
- *
- * MySQL variant - uses JSON instead of JSONB for tags
- */
-export const media = mysqlTable(
-  "media",
-  {
-    id: varchar("id", { length: 36 })
-      .primaryKey()
-      .$defaultFn(() => crypto.randomUUID()),
-
-    // File identification
-    filename: varchar("filename", { length: 255 }).notNull(),
-    originalFilename: varchar("original_filename", { length: 255 }).notNull(),
-    mimeType: varchar("mime_type", { length: 100 }).notNull(),
-    size: int("size").notNull(),
-
-    // Image/video dimensions
-    width: int("width"),
-    height: int("height"),
-    duration: int("duration"),
-
-    // Storage URLs
-    url: text("url").notNull(),
-    thumbnailUrl: text("thumbnail_url"),
-
-    // Crop point for smart image cropping (percentage from top-left, 0-100)
-    focalX: int("focal_x"),
-    focalY: int("focal_y"),
-
-    // Generated image size variants (JSON with size name → metadata)
-    sizes: json("sizes"),
-
-    // Metadata
-    altText: text("alt_text"),
-    caption: text("caption"),
-    tags: json("tags"), // MySQL uses JSON type for arrays
-
-    // Folder organization (null for root/unorganized files)
-    folderId: varchar("folder_id", { length: 255 }).references(
-      // The self-referencing FK requires an explicit return type to
-      // break TS's circular inference; AnyMySqlColumn matches what
-      // Drizzle's references() callback expects without the bare any.
-      (): AnyMySqlColumn => mediaFolders.id,
-      { onDelete: "set null" }
-    ),
-
-    // Ownership and timestamps
-    // Nullable: CLI seeds, data imports, and other system-context uploads
-    // may not have a user to attribute the upload to.
-    uploadedBy: varchar("uploaded_by", { length: 255 }).references(
-      () => users.id,
-      { onDelete: "cascade" }
-    ),
-    uploadedAt: datetime("uploaded_at").notNull().default(new Date()),
-    updatedAt: datetime("updated_at").notNull().default(new Date()),
-  },
-  t => [
-    index("media_uploaded_by_idx").on(t.uploadedBy),
-    index("media_mime_type_idx").on(t.mimeType),
-    index("media_uploaded_at_idx").on(t.uploadedAt),
-    index("media_folder_id_idx").on(t.folderId),
-  ]
-);
-
-/**
- * Media Folders table for organizing media files (MySQL)
- */
-export const mediaFolders = mysqlTable(
-  "media_folders",
-  {
-    id: varchar("id", { length: 255 })
-      .primaryKey()
-      .$defaultFn(() => crypto.randomUUID()),
-
-    // Folder information
-    name: varchar("name", { length: 255 }).notNull(),
-    description: text("description"),
-
-    // Hierarchy
-    parentId: varchar("parent_id", { length: 255 }).references(
-      // The self-referencing FK requires an explicit return type to
-      // break TS's circular inference; AnyMySqlColumn matches what
-      // Drizzle's references() callback expects without the bare any.
-      (): AnyMySqlColumn => mediaFolders.id,
-      { onDelete: "cascade" }
-    ),
-
-    // Ownership and timestamps
-    createdBy: varchar("created_by", { length: 191 })
-      .notNull()
-      .references(() => users.id, { onDelete: "cascade" }),
-    createdAt: datetime("created_at").notNull().default(new Date()),
-    updatedAt: datetime("updated_at").notNull().default(new Date()),
-  },
-  t => [
-    index("media_folders_parent_id_idx").on(t.parentId),
-    index("media_folders_created_by_idx").on(t.createdBy),
-    index("media_folders_created_at_idx").on(t.createdAt),
-  ]
-);
+// Media tables — moved to schemas/media/mysql.ts (Plan A Task 8).
+// Re-exported here so existing consumers and relations() blocks below keep working
+// until Task 16 sweeps imports to @nextly/schemas.
+export { media, mediaFolders, imageSizes } from "../../schemas/media/mysql";
+import { media, mediaFolders, imageSizes } from "../../schemas/media/mysql";
 
 export const mediaRelations = relations(media, ({ one }) => ({
   uploader: one(users, {
@@ -454,34 +354,8 @@ export const mediaFoldersRelations = relations(
   })
 );
 
-/**
- * Image Sizes table for named image size configurations (MySQL).
- * See postgres.ts for detailed documentation.
- */
-export const imageSizes = mysqlTable(
-  "image_sizes",
-  {
-    id: varchar("id", { length: 36 })
-      .primaryKey()
-      .$defaultFn(() => crypto.randomUUID()),
-
-    name: varchar("name", { length: 50 }).notNull(),
-    width: int("width"),
-    height: int("height"),
-    fit: varchar("fit", { length: 20 }).notNull().default("inside"),
-    quality: int("quality").notNull().default(80),
-    format: varchar("format", { length: 10 }).notNull().default("auto"),
-    isDefault: boolean("is_default").notNull().default(true),
-    sortOrder: int("sort_order").notNull().default(0),
-    createdAt: datetime("created_at")
-      .notNull()
-      .$defaultFn(() => new Date()),
-    updatedAt: datetime("updated_at")
-      .notNull()
-      .$defaultFn(() => new Date()),
-  },
-  t => [uniqueIndex("image_sizes_name_unique").on(t.name)]
-);
+// imageSizes moved to schemas/media/mysql.ts (Plan A Task 8) and re-exported
+// with the rest of the media tables above.
 
 // Metadata table for tracking dynamic collections
 export const dynamicCollections = mysqlTable(
